+++ .gitignore
@@ -21,3 +21,4 @@ logs
 # other
 eclipse
 run
+.vscode
