+++ build.gradle
@@ -1,149 +1,53 @@
-buildscript {
-	repositories {
-		jcenter()
-		mavenCentral()
-		maven { url "https://files.minecraftforge.net/maven" }
-		maven { url "https://plugins.gradle.org/m2/" }
-		//maven { url "https://dist.creeper.host/Sponge/maven" }
-	}
-	dependencies {
-		classpath group: 'net.minecraftforge.gradle', name: 'ForgeGradle', version: '3.+', changing: true
-		classpath 'gradle.plugin.com.matthewprenger:CurseGradle:1.1.0'
-		//classpath 'org.spongepowered:mixingradle:0.7-SNAPSHOT'
-	}
-}
-apply plugin: 'net.minecraftforge.gradle'
-apply plugin: 'eclipse'
-apply plugin: 'com.matthewprenger.cursegradle'
-apply plugin: 'maven'
-//apply plugin: 'org.spongepowered.mixin'
-
-apply from: 'https://files.latmod.com/public/markdown-git-changelog.gradle'
-
-def ENV = System.getenv()
-version = ENV.BUILD_NUMBER ? "${mod_version}.${ENV.BUILD_NUMBER}" : "${mod_version}.0"
-group = project.mod_package
-archivesBaseName = project.mod_name
-
-sourceCompatibility = targetCompatibility = compileJava.sourceCompatibility = compileJava.targetCompatibility = '1.8'
-
-minecraft {
-	mappings channel: 'official', version: '1.16.4'
-	//accessTransformer = file('src/main/resources/META-INF/accesstransformer.cfg')
-	runs {
-		client {
-			workingDirectory project.file('run')
-			//property 'forge.logging.markers', 'SCAN,REGISTRIES,REGISTRYDUMP'
-			property 'forge.logging.console.level', 'debug'
-			property 'mixin.env.disableRefMap', 'true'
-			//args '--mixin.config', 'kubejsui.mixins.json'
-
-			if (project.hasProperty('mc_dev_uuid')) args '--uuid', project.mc_dev_uuid
-			if (project.hasProperty('mc_dev_username')) args '--username', project.mc_dev_username
-			if (project.hasProperty('mc_dev_width')) args '--width', project.mc_dev_width
-			if (project.hasProperty('mc_dev_height')) args '--height', project.mc_dev_height
-
-			mods {
-				kubejs {
-					source sourceSets.main
-				}
-			}
-		}
-
-		server {
-			workingDirectory project.file('run')
-			//property 'forge.logging.markers', 'SCAN,REGISTRIES,REGISTRYDUMP'
-			property 'forge.logging.console.level', 'debug'
-			//args '--mixin.config', 'kubejsui.mixins.json'
-
-			args 'nogui'
-
-			mods {
-				kubejs {
-					source sourceSets.main
-				}
-			}
-		}
-
-		data {
-			workingDirectory project.file('run')
-			//property 'forge.logging.markers', 'SCAN,REGISTRIES,REGISTRYDUMP'
-			property 'forge.logging.console.level', 'debug'
-			//args '--mixin.config', 'kubejsui.mixins.json'
-
-			args '--mod', project.mod_id, '--all', '--output', file('src/generated/resources/')
-
-			mods {
-				kubejs {
-					source sourceSets.main
-				}
-			}
-		}
-	}
+plugins {
+    id "architectury-plugin" version "2.0.58"
+    id "forgified-fabric-loom" version "0.6.49" apply false
+    id "com.matthewprenger.cursegradle" version "1.4.0" apply false
 }
 
-repositories {
-	mavenLocal()
-	maven { url = "https://www.cursemaven.com" }
+architectury {
+    minecraft = rootProject.minecraft_version
 }
 
-dependencies {
-	minecraft "net.minecraftforge:forge:${forge_version}"
-	compile fg.deobf("curse.maven:rhino-416294:${rhino_file}")
-	compile fg.deobf("curse.maven:architectury-forge-419699:${architectury_file}")
-	compile fg.deobf("curse.maven:kubejs-238086:${kubejs_file}")
-	compile fg.deobf("curse.maven:jei-238222:${jei_file}")
+subprojects {
+    apply plugin: "forgified-fabric-loom"
+
+    loom {
+        silentMojangMappingsLicense()
+    }
 }
 
-jar {
-	manifest {
-		attributes([
-				//"MixinConfigs"            : "kubejsui.mixins.json",
-				"Specification-Title"     : project.mod_id,
-				"Specification-Vendor"    : project.mod_author,
-				"Specification-Version"   : "1",
-				"Implementation-Title"    : project.name,
-				"Implementation-Version"  : version,
-				"Implementation-Vendor"   : project.mod_author,
-				"Implementation-Timestamp": new Date().format("yyyy-MM-dd'T'HH:mm:ssZ")
-		])
-	}
-}
+allprojects {
+    apply plugin: "java"
+    apply plugin: "architectury-plugin"
+    apply from: "https://files.latmod.com/public/markdown-git-changelog.gradle"
 
-def reobfFile = file("$buildDir/reobfJar/output.jar")
-def reobfArtifact = artifacts.add('default', reobfFile) {
-	type 'jar'
-	builtBy 'reobfJar'
-}
+    archivesBaseName = rootProject.archives_base_name
+    def ENV = System.getenv()
+    version = ENV.BUILD_NUMBER ? "${rootProject.mod_version}.${ENV.BUILD_NUMBER}" : "${rootProject.mod_version}.custom"
+    group = rootProject.maven_group
 
-if (ENV.LOCAL_MAVEN) {
-	uploadArchives {
-		repositories {
-			mavenDeployer {
-				repository(url: "file://${ENV.LOCAL_MAVEN}")
-			}
-		}
-	}
-}
+    tasks.withType(JavaCompile) {
+        options.encoding = "UTF-8"
 
-if (ENV.CURSEFORGE_KEY) {
-	curseforge {
-		apiKey = ENV.CURSEFORGE_KEY
-		project {
-			id = project.curseforge_id
-			releaseType = "release"
-			addGameVersion "1.16.2"
-			addGameVersion "1.16.3"
-			addGameVersion "1.16.4"
-			relations {
-				requiredDependency 'kubejs'
-			}
-			changelog = getGitChangelog
-			changelogType = 'markdown'
-		}
-	}
-}
+        // The Minecraft launcher currently installs Java 8 for users, so your mod probably wants to target Java 8 too
+        // JDK 9 introduced a new way of specifying this that will make sure no newer classes or methods are used.
+        // We'll use that if it's available, but otherwise we'll use the older option.
+        def targetVersion = 8
+        if (JavaVersion.current().isJava9Compatible()) {
+            options.release = targetVersion
+        }
+    }
 
-//mixin {
-//	add sourceSets.main, "kubejsui.refmap.json"
-//}
+    java {
+        withSourcesJar()
+    }
+
+    repositories {
+        maven {
+            url "https://www.cursemaven.com"
+            content {
+                includeGroup "curse.maven"
+            }
+        }
+    }
+}
