+++ common/build.gradle
@@ -1,0 +1,14 @@
+dependencies {
+    minecraft "com.mojang:minecraft:${rootProject.minecraft_version}"
+    mappings minecraft.officialMojangMappings()
+
+    modCompile "me.shedaniel:architectury:${rootProject.architectury_version}"
+    compileOnly "com.google.code.findbugs:jsr305:3.+"
+    // We depend on fabric loader here to use the fabric @Environment annotations
+    // Do NOT use other classes from fabric loader
+    modCompile "net.fabricmc:fabric-loader:${rootProject.fabric_loader_version}"
+}
+
+architectury {
+    common()
+}
