+++ common/src/main/java/dev/latvian/kubejs/ui/widget/Mouse.java
