+++ gradle.properties
@@ -1,13 +1,30 @@
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
+
+minecraft_version=1.16.4
+
 mod_id=kubejs_ui
-mod_name=kubejs-ui
-mod_package=dev.latvian.mods
+archives_base_name=kubejs-ui
 mod_version=1604.1.0
+maven_group=dev.latvian.mods
 mod_author=LatvianModder
-curseforge_id=423490
-forge_version=1.16.4-35.0.7
-rhino_file=3124036
-architectury_file=3128224
-kubejs_file=3137415
-jei_file=3136600
+
+architectury_version=1.3.84
+
+curseforge_id_forge=423490
+curseforge_id_fabric=0
+
+rhino_file=3162251
+
+# fabric
+fabric_loader_version=0.11.1
+fabric_api_version=0.29.3+1.16
+
+fabric_kjs_file=3102076
+fabric_rei_file=3175580
+
+# forge
+forge_version=35.0.7
+
+forge_kjs_file=3137415
+forge_jei_file=3136600
