--- src/main/java/dev/latvian/kubejs/ui/KubeJSUI.java
@@ -1,12 +1,0 @@
-package dev.latvian.kubejs.ui;
-
-import net.minecraftforge.fml.common.Mod;
-
-/**
- * @author LatvianModder
- */
-@Mod(KubeJSUI.MOD_ID)
-public class KubeJSUI
-{
-	public static final String MOD_ID = "kubejs_ui";
-}
