--- src/main/resources/META-INF/mods.toml
@@ -1,21 +1,0 @@
-modLoader = "javafml"
-loaderVersion = "[32,)"
-issueTrackerURL = "https://github.com/KubeJS-Mods/KubeJS-UI/issues"
-license = "GNU LGPLv3"
-
-[[mods]]
-modId = "kubejs_ui"
-version = "${file.jarVersion}"
-displayName = "KubeJS UI"
-authors = "LatvianModder"
-description = '''
-KubeJS UI - Edit main menu with scripts
-'''
-license = "GNU LGPLv3"
-
-[[dependencies.kubejs_ui]]
-modId = "kubejs"
-mandatory = true
-versionRange = "[1604,)"
-ordering = "AFTER"
-side = "BOTH"
