--- src/main/resources/assets/forge/kubejsui.json
@@ -1,5 +1,0 @@
-{
-	"actions": {
-		"mod_list": "dev.latvian.kubejs.ui.VanillaActions.FORGE_MOD_LIST"
-	}
-}
