--- src/main/resources/assets/kubejsui/kubejsui.json
@@ -1,5 +1,0 @@
-{
-	"actions": {
-		"toggle_shaders": "dev.latvian.kubejs.ui.VanillaActions.KUBEJSUI_TOGGLE_SHADERS"
-	}
-}
