--- src/main/resources/assets/kubejsui/shaders/program/none.vsh
@@ -1,6 +1,0 @@
-#version 110
-attribute vec3 position;
-
-void main() {
-    gl_Position = vec4(position, 1.0);
-}
