--- src/main/resources/assets/minecraft/kubejsui.json
@@ -1,14 +1,0 @@
-{
-	"screens": {
-		"net.minecraft.client.gui.screen.MainMenuScreen": "main_menu"
-	},
-	"actions": {
-		"singleplayer": "dev.latvian.kubejs.ui.VanillaActions.SINGLEPLAYER",
-		"multiplayer": "dev.latvian.kubejs.ui.VanillaActions.MULTIPLAYER",
-		"realms": "dev.latvian.kubejs.ui.VanillaActions.REALMS",
-		"language": "dev.latvian.kubejs.ui.VanillaActions.LANGUAGE",
-		"options": "dev.latvian.kubejs.ui.VanillaActions.OPTIONS",
-		"quit": "dev.latvian.kubejs.ui.VanillaActions.QUIT",
-		"accessibility": "dev.latvian.kubejs.ui.VanillaActions.ACCESSIBILITY"
-	}
-}
