--- src/main/resources/pack.mcmeta
@@ -1,6 +1,0 @@
-{
-	"pack": {
-		"description": "KubeJS UI Resources",
-		"pack_format": 6
-	}
-}
